- Start Users section - Deal with language defaults and missing content

+++ module/Admin/src/Admin/Controller/LanguageController.php
@@ -4,6 +4,7 @@ namespace Admin\Controller;
 
 
 use Admin\Form\Language as LanguageForm;
+use Application\Model\Entity\Category;
 use Application\Model\Entity\Lang;
 use Zend\Mvc\Controller\AbstractActionController;
 use Zend\i18n\Translator\Translator;
@@ -70,6 +71,7 @@ class LanguageController extends AbstractActionController
         }else{
             $language = new Lang();
         }
+        $oldDefaultLanguage = $entityManager->getRepository(get_class($language))->findOneByStatus(Lang::STATUS_DEFAULT);
 
         $form = new LanguageForm();
         $form->bind($language);
@@ -85,15 +87,71 @@ class LanguageController extends AbstractActionController
                 $this->getRequest()->getPost()->toArray(),
                 $this->getRequest()->getFiles()->toArray()
             );
+            if($language->isDefault()) unset($post['status']);//ensures no status can be changed if lang is default
+
             $form->setData($post);
-            if($form->isValid($post['isoCode'], $oldIso)){
-                //if this is the new default language, change the old default to status active
-                if($oldStatus != $post['status'] && $post['status'] == 2){
-                    $oldDefaultLanguage = $entityManager->getRepository(get_class($language))->findOneByStatus(2);
+            if($form->isValid($post['isoCode'], $oldIso, $language->isDefault())){
+                //if this is the new default language, change the old default to status active, and populate the missing content in the new default lang
+                if(isset($post['status']) && $oldStatus != $post['status'] && $language->isDefault($post['status'])){
+
+                    //region fill missing content in categories
+                    $oldDefaultLanguageId = $oldDefaultLanguage->getId();
+                    $editedLanguageId = $language->getId();//note! this is null if new language
+                    $getMissingData = function($contentCollection)use($entityManager,$language,$oldDefaultLanguageId,$editedLanguageId){
+                        $defaultContent = null;
+                        $editedContent = null;
+                        foreach($contentCollection as $content){
+                            $contentLangId = $content->getLang()->getId();
+                            if($contentLangId == $oldDefaultLanguageId){
+                                $defaultContent = $content;
+                                continue;
+                            }
+                            if($editedLanguageId && $contentLangId == $editedLanguageId){//if($editedLanguageId) = do this only in case of action == edit
+                                $editedContent = $content;
+                                continue;
+                            }
+                            if($defaultContent && $editedContent) break;//job done
+                        }
+                        return [
+                            'default_content' => $defaultContent,
+                            'edited_content' => $editedContent
+                        ];
+                    };
+                    //parse through all the categories
+                    $categories = $entityManager->getRepository(get_class(new Category()))->findByType(1);
+                    foreach($categories as $category){
+                        //region copy category content
+                        $categoryContents = $getMissingData($category->getContent());
+                        if(!$categoryContents['edited_content'] && $categoryContents['default_content']){
+                            $newContent = clone $categoryContents['default_content'];
+                            $newContent->setLang($language);
+                            $category->addCategoryContent($newContent);
+                            $entityManager->persist($category);
+                        }
+                        //endregion
+
+                        //region copy listings content
+                        foreach($category->getListings() as $listing){
+                            //deal with content
+                            $listingContents = $getMissingData($listing->getContent());
+                            if(!$listingContents['edited_content'] && $listingContents['default_content']){
+                                $newListingContent = clone $listingContents['default_content'];
+                                $newListingContent->setLang($language);
+                                $listing->addContent($newListingContent);
+                                $entityManager->persist($listing);
+                            }
+                        }
+                        //endregion
+                    }
+                    //endregion
+
+
                     if($oldDefaultLanguage instanceof Lang){
-                        $oldDefaultLanguage->setStatus(1);
+                        $oldDefaultLanguage->setStatus(Lang::STATUS_ACTIVE);
                         $entityManager->persist($oldDefaultLanguage);
                     }
+
+                    //v_todo - if the language is edited and made inactive, purge all it's content
                 }
 
                 $entityManager->persist($language);
@@ -135,6 +193,10 @@ class LanguageController extends AbstractActionController
 
         $entityManager = $this->getServiceLocator()->get('entity-manager');
         $lang = $entityManager->find(get_class(new Lang()), $id);
+        if($lang->isDefault()){
+            $this->flashMessenger()->addErrorMessage($this->translator->translate('The default language cannot be deleted'));
+            return $this->redir()->toRoute('admin/default', ['controller' => 'language', 'page' => $page]);
+        }
         if($lang instanceof Lang){
             $entityManager->remove($lang);
             $entityManager->flush();
